apps/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
pandas
pydantic
python-multipart
boto3
botocore[crt]
numpy
scipy
scikit-learn